Refactor project into modular packages
