+++ utils/__init__.py
@@ -1,0 +1,6 @@
+from .image import cvimg_to_qpixmap, encode_png_bytes, decode_png_bytes, Matcher, MatchResult
+from .helpers import (
+    info, warn, err,
+    _normalize_point_result, _InlinePointOverlay, safe_select_point,
+    make_letter_icon, hk_normalize, hk_pretty, hk_to_tuple,
+)
